refactor: split app scenario tests
