--- internal/app/app_test.go
@@ -1,1011 +1,0 @@
-package app
-
-import (
-	"bytes"
-	"context"
-	"errors"
-	"os"
-	"path/filepath"
-	"strings"
-	"testing"
-	"time"
-
-	"github.com/ben-ranford/lopper/internal/analysis"
-	"github.com/ben-ranford/lopper/internal/notify"
-	"github.com/ben-ranford/lopper/internal/report"
-	"github.com/ben-ranford/lopper/internal/thresholds"
-	"github.com/ben-ranford/lopper/internal/ui"
-)
-
-const (
-	testSnapshotPath        = "snapshot.txt"
-	testBaselinePath        = "baseline.json"
-	missingBaselineFileName = "missing.json"
-	saveBaselineStoreErr    = "--save-baseline requires --baseline-store"
-	baselineStorePath       = ".artifacts/baselines"
-	executeAnalyseErrFmt    = "execute analyse: %v"
-	deniedLicenseSPDX       = "GPL-3.0-ONLY"
-	testRuntimeTracePath    = "trace.ndjson"
-)
-
-type fakeAnalyzer struct {
-	report  report.Report
-	err     error
-	lastReq analysis.Request
-	called  bool
-}
-
-func (f *fakeAnalyzer) Analyse(_ context.Context, req analysis.Request) (report.Report, error) {
-	f.called = true
-	f.lastReq = req
-	return f.report, f.err
-}
-
-type fakeTUI struct {
-	startErr       error
-	snapshotErr    error
-	startCalled    bool
-	snapshotCalled bool
-	lastOptions    ui.Options
-	lastSnapshot   string
-}
-
-type fakeNotifier struct {
-	lastDelivery notify.Delivery
-	called       bool
-	err          error
-}
-
-func (f *fakeNotifier) Notify(_ context.Context, delivery notify.Delivery) error {
-	f.called = true
-	f.lastDelivery = delivery
-	return f.err
-}
-
-func (f *fakeTUI) Start(_ context.Context, opts ui.Options) error {
-	f.startCalled = true
-	f.lastOptions = opts
-	return f.startErr
-}
-
-func (f *fakeTUI) Snapshot(_ context.Context, opts ui.Options, outputPath string) error {
-	f.snapshotCalled = true
-	f.lastOptions = opts
-	f.lastSnapshot = outputPath
-	return f.snapshotErr
-}
-
-func assertContainsAll(t *testing.T, output string, expected []string) {
-	t.Helper()
-	for _, value := range expected {
-		if !strings.Contains(output, value) {
-			t.Fatalf("expected output to include %q", value)
-		}
-	}
-}
-
-func assertForwardedAnalyseRequest(t *testing.T, got analysis.Request) {
-	t.Helper()
-	checks := []struct {
-		name string
-		ok   bool
-	}{
-		{"low confidence threshold", got.LowConfidenceWarningPercent != nil && *got.LowConfidenceWarningPercent == 33},
-		{"min usage threshold", got.MinUsagePercentForRecommendations != nil && *got.MinUsagePercentForRecommendations == 44},
-		{"runtime profile", got.RuntimeProfile == "browser-import"},
-		{"scope mode", got.ScopeMode == ScopeModeChangedPackages},
-		{"cache options", got.Cache != nil && !got.Cache.Enabled && got.Cache.Path == "/tmp/lopper-cache" && got.Cache.ReadOnly},
-		{"suggest only", got.SuggestOnly},
-		{"removal candidate weights", got.RemovalCandidateWeights != nil && got.RemovalCandidateWeights.Usage == 0.6 && got.RemovalCandidateWeights.Impact == 0.2 && got.RemovalCandidateWeights.Confidence == 0.2},
-	}
-	for _, check := range checks {
-		if !check.ok {
-			t.Fatalf("expected forwarded analyse request field: %s (got=%#v)", check.name, got)
-		}
-	}
-}
-
-func TestExecuteAnalyseEmitsEffectiveThresholds(t *testing.T) {
-	analyzer := &fakeAnalyzer{
-		report: report.Report{
-			RepoPath: ".",
-			Dependencies: []report.DependencyReport{
-				{Name: "lodash", UsedExportsCount: 1, TotalExportsCount: 2, UsedPercent: 50},
-			},
-		},
-	}
-	application := &App{Analyzer: analyzer, Formatter: report.NewFormatter()}
-
-	req := DefaultRequest()
-	req.Mode = ModeAnalyse
-	req.Analyse.TopN = 1
-	req.Analyse.ScopeMode = ScopeModeChangedPackages
-	req.Analyse.Format = report.FormatJSON
-	req.Analyse.SuggestOnly = true
-	req.Analyse.RuntimeProfile = "browser-import"
-	req.Analyse.CacheEnabled = false
-	req.Analyse.CachePath = "/tmp/lopper-cache"
-	req.Analyse.CacheReadOnly = true
-	req.Analyse.PolicySources = []string{"cli", "defaults"}
-	req.Analyse.Thresholds = thresholds.Values{
-		FailOnIncreasePercent:             0,
-		LowConfidenceWarningPercent:       33,
-		MinUsagePercentForRecommendations: 44,
-		RemovalCandidateWeightUsage:       0.6,
-		RemovalCandidateWeightImpact:      0.2,
-		RemovalCandidateWeightConfidence:  0.2,
-	}
-
-	output, err := application.Execute(context.Background(), req)
-	if err != nil {
-		t.Fatalf(executeAnalyseErrFmt, err)
-	}
-	assertContainsAll(t, output, []string{`"effectiveThresholds"`, `"effectivePolicy"`, `"sources": [`, `"cli"`, `"lowConfidenceWarningPercent": 33`})
-	assertForwardedAnalyseRequest(t, analyzer.lastReq)
-}
-
-func TestNewApp(t *testing.T) {
-	appInstance := New(&bytes.Buffer{}, strings.NewReader(""))
-	if appInstance == nil {
-		t.Fatalf("expected app instance")
-	}
-}
-
-func TestExecuteAnalyseFailOnIncreaseThreshold(t *testing.T) {
-	delta := 3.5
-	analyzer := &fakeAnalyzer{
-		report: report.Report{
-			RepoPath:             ".",
-			Dependencies:         []report.DependencyReport{{Name: "lodash", UsedExportsCount: 1, TotalExportsCount: 2, UsedPercent: 50}},
-			WasteIncreasePercent: &delta,
-		},
-	}
-	application := &App{Analyzer: analyzer, Formatter: report.NewFormatter()}
-
-	req := DefaultRequest()
-	req.Mode = ModeAnalyse
-	req.Analyse.TopN = 1
-	req.Analyse.Thresholds = thresholds.Values{
-		FailOnIncreasePercent:             2,
-		LowConfidenceWarningPercent:       thresholds.DefaultLowConfidenceWarningPercent,
-		MinUsagePercentForRecommendations: thresholds.DefaultMinUsagePercentForRecommendations,
-	}
-
-	_, err := application.Execute(context.Background(), req)
-	if err == nil {
-		t.Fatalf("expected fail-on-increase error")
-	}
-	if !errors.Is(err, ErrFailOnIncrease) {
-		t.Fatalf("expected ErrFailOnIncrease, got %v", err)
-	}
-}
-
-func TestExecuteAnalyseDispatchesNotifications(t *testing.T) {
-	analyzer := &fakeAnalyzer{
-		report: report.Report{
-			RepoPath:      ".",
-			Dependencies:  []report.DependencyReport{{Name: "lodash", UsedExportsCount: 1, TotalExportsCount: 2, UsedPercent: 50}},
-			Warnings:      []string{"existing warning"},
-			GeneratedAt:   time.Now().UTC(),
-			SchemaVersion: report.SchemaVersion,
-		},
-	}
-	slackNotifier := &fakeNotifier{err: errors.New("slack unreachable")}
-	application := &App{
-		Analyzer:  analyzer,
-		Formatter: report.NewFormatter(),
-		Notify: notify.NewDispatcher(map[notify.Channel]notify.Notifier{
-			notify.ChannelSlack: slackNotifier,
-		}),
-	}
-
-	req := DefaultRequest()
-	req.Mode = ModeAnalyse
-	req.Analyse.TopN = 1
-	req.Analyse.Format = report.FormatJSON
-	req.Analyse.Notifications = notify.Config{
-		Slack: notify.ChannelConfig{
-			WebhookURL: "https://hooks.slack.com/services/A/B/SECRET",
-			Trigger:    notify.TriggerAlways,
-		},
-	}
-
-	output, err := application.Execute(context.Background(), req)
-	if err != nil {
-		t.Fatalf(executeAnalyseErrFmt, err)
-	}
-	if !slackNotifier.called {
-		t.Fatalf("expected notification dispatch call")
-	}
-	if slackNotifier.lastDelivery.Channel != notify.ChannelSlack {
-		t.Fatalf("expected slack delivery channel, got %q", slackNotifier.lastDelivery.Channel)
-	}
-	assertContainsAll(t, output, []string{
-		"existing warning",
-		"notification delivery failed for slack",
-		"https://hooks.slack.com",
-	})
-	if strings.Contains(output, "SECRET") {
-		t.Fatalf("expected redacted webhook warning, got %q", output)
-	}
-}
-
-func TestExecuteTUIStartAndSnapshot(t *testing.T) {
-	tui := &fakeTUI{}
-	application := &App{
-		Analyzer:  &fakeAnalyzer{},
-		Formatter: report.NewFormatter(),
-		TUI:       tui,
-	}
-
-	req := DefaultRequest()
-	req.Mode = ModeTUI
-	req.TUI.TopN = 5
-	req.TUI.Language = "all"
-	req.TUI.Filter = "lod"
-	req.TUI.Sort = "name"
-	req.TUI.PageSize = 3
-
-	if _, err := application.Execute(context.Background(), req); err != nil {
-		t.Fatalf("execute tui start: %v", err)
-	}
-	if !tui.startCalled || tui.snapshotCalled {
-		t.Fatalf("expected Start to be called only once")
-	}
-
-	req.TUI.SnapshotPath = testSnapshotPath
-	if _, err := application.Execute(context.Background(), req); err != nil {
-		t.Fatalf("execute tui snapshot: %v", err)
-	}
-	if !tui.snapshotCalled || tui.lastSnapshot != testSnapshotPath {
-		t.Fatalf("expected Snapshot call with output path, got called=%v path=%q", tui.snapshotCalled, tui.lastSnapshot)
-	}
-}
-
-func TestExecuteUnknownMode(t *testing.T) {
-	application := &App{Analyzer: &fakeAnalyzer{}, Formatter: report.NewFormatter()}
-	_, err := application.Execute(context.Background(), Request{Mode: "unknown"})
-	if !errors.Is(err, ErrUnknownMode) {
-		t.Fatalf("expected ErrUnknownMode, got %v", err)
-	}
-}
-
-func TestApplyBaselineIfNeededWithBaselineFile(t *testing.T) {
-	tmp := t.TempDir()
-	path := filepath.Join(tmp, testBaselinePath)
-	data := `{"schemaVersion":"0.1.0","generatedAt":"2026-01-01T00:00:00Z","repoPath":".","dependencies":[{"name":"dep","usedExportsCount":5,"totalExportsCount":10,"usedPercent":50,"estimatedUnusedBytes":0}]}` + "\n"
-	if err := os.WriteFile(path, []byte(data), 0o600); err != nil {
-		t.Fatalf("write baseline: %v", err)
-	}
-
-	application := &App{Formatter: report.NewFormatter()}
-	current := report.Report{
-		Dependencies: []report.DependencyReport{
-			{Name: "dep", UsedExportsCount: 4, TotalExportsCount: 10, UsedPercent: 40},
-		},
-	}
-	updated, err := application.applyBaselineIfNeeded(current, ".", AnalyseRequest{BaselinePath: path, Format: report.FormatJSON})
-	if err != nil {
-		t.Fatalf("apply baseline: %v", err)
-	}
-	if updated.WasteIncreasePercent == nil {
-		t.Fatalf("expected waste increase to be computed")
-	}
-	if updated.BaselineComparison == nil {
-		t.Fatalf("expected baseline comparison details to be present")
-	}
-}
-
-func TestValidateFailOnIncreaseRequiresBaseline(t *testing.T) {
-	err := validateFailOnIncrease(report.Report{}, 2)
-	if !errors.Is(err, ErrBaselineRequired) {
-		t.Fatalf("expected ErrBaselineRequired, got %v", err)
-	}
-	if err := validateFailOnIncrease(report.Report{}, 0); err != nil {
-		t.Fatalf("expected no error when threshold disabled, got %v", err)
-	}
-}
-
-func TestValidateDeniedLicenses(t *testing.T) {
-	reportData := report.Report{
-		Dependencies: []report.DependencyReport{
-			{Name: "a", License: &report.DependencyLicense{SPDX: "MIT", Denied: false}},
-			{Name: "b", License: &report.DependencyLicense{SPDX: deniedLicenseSPDX, Denied: true}},
-		},
-	}
-	if err := validateDeniedLicenses(reportData, true); !errors.Is(err, ErrDeniedLicenses) {
-		t.Fatalf("expected denied license error, got %v", err)
-	}
-	if err := validateDeniedLicenses(reportData, false); err != nil {
-		t.Fatalf("expected no error when policy disabled, got %v", err)
-	}
-}
-
-func TestValidateDeniedLicensesNoDeniedReturnsNil(t *testing.T) {
-	reportData := report.Report{
-		Dependencies: []report.DependencyReport{
-			{Name: "a", License: &report.DependencyLicense{SPDX: "MIT", Denied: false}},
-		},
-	}
-	if err := validateDeniedLicenses(reportData, true); err != nil {
-		t.Fatalf("expected no denied license error, got %v", err)
-	}
-}
-
-func TestValidateDeniedLicensesBaselineNewDeniedBranch(t *testing.T) {
-	reportData := report.Report{
-		Dependencies: []report.DependencyReport{
-			{Name: "safe", License: &report.DependencyLicense{SPDX: "MIT", Denied: false}},
-		},
-		BaselineComparison: &report.BaselineComparison{
-			NewDeniedLicenses: []report.DeniedLicenseDelta{
-				{Name: "unsafe", Language: "js-ts", SPDX: deniedLicenseSPDX},
-			},
-		},
-	}
-	if err := validateDeniedLicenses(reportData, true); !errors.Is(err, ErrDeniedLicenses) {
-		t.Fatalf("expected denied license error from baseline new-denied branch, got %v", err)
-	}
-
-	reportData.BaselineComparison.NewDeniedLicenses = nil
-	reportData.Dependencies = []report.DependencyReport{
-		{Name: "existing-denied", License: &report.DependencyLicense{SPDX: deniedLicenseSPDX, Denied: true}},
-	}
-	if err := validateDeniedLicenses(reportData, true); err != nil {
-		t.Fatalf("expected no denied-license error for baseline mode without newly introduced denied licenses, got %v", err)
-	}
-}
-
-func TestValidateUncertaintyThreshold(t *testing.T) {
-	reportData := report.Report{
-		UsageUncertainty: &report.UsageUncertainty{
-			UncertainImportUses: 2,
-		},
-	}
-	if err := validateUncertaintyThreshold(reportData, 2); err != nil {
-		t.Fatalf("expected no uncertainty threshold error at boundary, got %v", err)
-	}
-	if err := validateUncertaintyThreshold(reportData, 1); !errors.Is(err, ErrUncertaintyThresholdExceeded) {
-		t.Fatalf("expected uncertainty threshold error, got %v", err)
-	}
-}
-
-func TestExecuteAnalyseUncertaintyThresholdError(t *testing.T) {
-	analyzer := &fakeAnalyzer{
-		report: report.Report{
-			RepoPath:     ".",
-			Dependencies: []report.DependencyReport{{Name: "lodash", UsedExportsCount: 1, TotalExportsCount: 2, UsedPercent: 50}},
-			UsageUncertainty: &report.UsageUncertainty{
-				UncertainImportUses: 2,
-			},
-		},
-	}
-	application := &App{Analyzer: analyzer, Formatter: report.NewFormatter()}
-
-	req := DefaultRequest()
-	req.Mode = ModeAnalyse
-	req.Analyse.TopN = 1
-	req.Analyse.Format = report.FormatJSON
-	req.Analyse.Thresholds.MaxUncertainImportCount = 1
-
-	output, err := application.Execute(context.Background(), req)
-	if !errors.Is(err, ErrUncertaintyThresholdExceeded) {
-		t.Fatalf("expected uncertainty threshold error, got %v", err)
-	}
-	if !strings.Contains(output, `"effectiveThresholds"`) {
-		t.Fatalf("expected formatted output on threshold failure, got %q", output)
-	}
-}
-
-func TestExecuteAnalyseDeniedLicensesError(t *testing.T) {
-	analyzer := &fakeAnalyzer{
-		report: report.Report{
-			RepoPath: ".",
-			Dependencies: []report.DependencyReport{
-				{
-					Name:    "copyleft",
-					License: &report.DependencyLicense{SPDX: deniedLicenseSPDX, Denied: true},
-				},
-			},
-		},
-	}
-	application := &App{Analyzer: analyzer, Formatter: report.NewFormatter()}
-
-	req := DefaultRequest()
-	req.Mode = ModeAnalyse
-	req.Analyse.TopN = 1
-	req.Analyse.Format = report.FormatJSON
-	req.Analyse.Thresholds.LicenseFailOnDeny = true
-
-	output, err := application.Execute(context.Background(), req)
-	if !errors.Is(err, ErrDeniedLicenses) {
-		t.Fatalf("expected denied licenses error, got %v", err)
-	}
-	if !strings.Contains(output, `"effectivePolicy"`) {
-		t.Fatalf("expected formatted output on denied-license failure, got %q", output)
-	}
-}
-
-func TestExecuteAnalyseAnalyzerError(t *testing.T) {
-	expected := errors.New("analyse failed")
-	application := &App{
-		Analyzer:  &fakeAnalyzer{err: expected},
-		Formatter: report.NewFormatter(),
-	}
-	req := DefaultRequest()
-	req.Mode = ModeAnalyse
-	req.Analyse.Dependency = "lodash"
-	_, err := application.Execute(context.Background(), req)
-	if !errors.Is(err, expected) {
-		t.Fatalf("expected analyzer error, got %v", err)
-	}
-}
-
-func TestApplyBaselineIfNeededFormatAndLoadErrors(t *testing.T) {
-	application := &App{Formatter: report.NewFormatter()}
-
-	_, err := application.applyBaselineIfNeeded(report.Report{}, ".", AnalyseRequest{
-		Format:       report.FormatJSON,
-		BaselinePath: filepath.Join(t.TempDir(), missingBaselineFileName),
-	})
-	if err == nil {
-		t.Fatalf("expected missing baseline load error")
-	}
-
-	_, err = application.applyBaselineIfNeeded(report.Report{}, ".", AnalyseRequest{
-		Format:            report.FormatJSON,
-		BaselineStorePath: filepath.Join(t.TempDir(), "baselines"),
-	})
-	if err == nil {
-		t.Fatalf("expected baseline-store comparison error without key/commit")
-	}
-}
-
-func TestValidateFailOnIncreaseAllowsWithinThreshold(t *testing.T) {
-	delta := 2.0
-	err := validateFailOnIncrease(report.Report{WasteIncreasePercent: &delta}, 2)
-	if err != nil {
-		t.Fatalf("expected no error at threshold boundary, got %v", err)
-	}
-}
-
-func TestExecuteTUIPropagatesErrors(t *testing.T) {
-	tui := &fakeTUI{startErr: errors.New("start failed")}
-	application := &App{
-		Analyzer:  &fakeAnalyzer{},
-		Formatter: report.NewFormatter(),
-		TUI:       tui,
-	}
-	req := DefaultRequest()
-	req.Mode = ModeTUI
-	if _, err := application.Execute(context.Background(), req); err == nil {
-		t.Fatalf("expected start error")
-	}
-
-	tui = &fakeTUI{snapshotErr: errors.New("snapshot failed")}
-	application.TUI = tui
-	req.TUI.SnapshotPath = testSnapshotPath
-	if _, err := application.Execute(context.Background(), req); err == nil {
-		t.Fatalf("expected snapshot error")
-	}
-}
-
-func TestExecuteAnalyseBaselineAndApplyBaselineErrors(t *testing.T) {
-	analyzer := &fakeAnalyzer{
-		report: report.Report{
-			Dependencies: []report.DependencyReport{
-				{Name: "dep", UsedExportsCount: 1, TotalExportsCount: 2, UsedPercent: 50},
-			},
-		},
-	}
-	application := &App{Analyzer: analyzer, Formatter: report.NewFormatter()}
-
-	req := DefaultRequest()
-	req.Mode = ModeAnalyse
-	req.Analyse.Dependency = "dep"
-	req.Analyse.Format = report.FormatJSON
-	req.Analyse.BaselinePath = filepath.Join(t.TempDir(), missingBaselineFileName)
-	if _, err := application.Execute(context.Background(), req); err == nil {
-		t.Fatalf("expected execute analyse error when baseline path is missing")
-	}
-
-	tmp := t.TempDir()
-	baselinePath := filepath.Join(tmp, testBaselinePath)
-	content := `{"schemaVersion":"0.1.0","generatedAt":"2026-01-01T00:00:00Z","repoPath":".","dependencies":[{"name":"dep","usedExportsCount":0,"totalExportsCount":0,"usedPercent":0}]}` + "\n"
-	if err := os.WriteFile(baselinePath, []byte(content), 0o600); err != nil {
-		t.Fatalf("write baseline file: %v", err)
-	}
-	current := report.Report{
-		Dependencies: []report.DependencyReport{{Name: "dep", UsedExportsCount: 1, TotalExportsCount: 2, UsedPercent: 50}},
-	}
-	_, err := application.applyBaselineIfNeeded(current, ".", AnalyseRequest{BaselinePath: baselinePath, Format: report.FormatJSON})
-	if err == nil {
-		t.Fatalf("expected baseline application error for zero baseline totals")
-	}
-}
-
-func TestSaveBaselineIfNeeded(t *testing.T) {
-	application := &App{Formatter: report.NewFormatter()}
-	base := report.Report{
-		SchemaVersion: "0.1.0",
-		RepoPath:      ".",
-		Dependencies: []report.DependencyReport{
-			{Name: "dep", UsedExportsCount: 1, TotalExportsCount: 2, UsedPercent: 50},
-		},
-	}
-	dir := t.TempDir()
-	now := testTime()
-
-	saveReq := AnalyseRequest{
-		SaveBaseline:      true,
-		BaselineStorePath: dir,
-		BaselineLabel:     "nightly",
-	}
-	updated, err := application.saveBaselineIfNeeded(base, ".", saveReq, now)
-	if err != nil {
-		t.Fatalf("save baseline: %v", err)
-	}
-	if len(updated.Warnings) == 0 || !strings.Contains(updated.Warnings[0], "saved immutable baseline snapshot:") {
-		t.Fatalf("expected save warning, got %#v", updated.Warnings)
-	}
-}
-
-func TestSaveBaselineIfNeededRequiresStorePath(t *testing.T) {
-	application := &App{Formatter: report.NewFormatter()}
-	_, err := application.saveBaselineIfNeeded(report.Report{}, ".", AnalyseRequest{SaveBaseline: true}, testTime())
-	if err == nil || !strings.Contains(err.Error(), saveBaselineStoreErr) {
-		t.Fatalf("expected missing baseline-store error, got %v", err)
-	}
-}
-
-func TestSaveBaselineIfNeededResolveSaveBaselineKeyError(t *testing.T) {
-	application := &App{Formatter: report.NewFormatter()}
-	nonRepo := filepath.Join(t.TempDir(), "nonexistent", "repo")
-	saveReq := AnalyseRequest{
-		SaveBaseline:      true,
-		BaselineStorePath: t.TempDir(),
-	}
-	_, err := application.saveBaselineIfNeeded(report.Report{}, nonRepo, saveReq, testTime())
-	if err == nil || !strings.Contains(err.Error(), "unable to resolve git commit") {
-		t.Fatalf("expected save-baseline key resolution error, got %v", err)
-	}
-}
-
-func TestResolveSaveBaselineKeyBranches(t *testing.T) {
-	if key, err := resolveSaveBaselineKey(".", AnalyseRequest{BaselineLabel: "nightly"}); err != nil || key != "label:nightly" {
-		t.Fatalf("expected label-based key, got key=%q err=%v", key, err)
-	}
-	if key, err := resolveSaveBaselineKey(".", AnalyseRequest{BaselineKey: "commit:abc"}); err != nil || key != "commit:abc" {
-		t.Fatalf("expected explicit key, got key=%q err=%v", key, err)
-	}
-	if key, err := resolveSaveBaselineKey(".", AnalyseRequest{}); err != nil || !strings.HasPrefix(key, "commit:") {
-		t.Fatalf("expected commit-derived key, got key=%q err=%v", key, err)
-	}
-
-	nonRepo := filepath.Join(t.TempDir(), "nonexistent", "repo")
-	if _, err := resolveSaveBaselineKey(nonRepo, AnalyseRequest{}); err == nil || !strings.Contains(err.Error(), "unable to resolve git commit") {
-		t.Fatalf("expected missing git key resolution error, got %v", err)
-	}
-}
-
-func TestResolveBaselineComparisonPathsBranches(t *testing.T) {
-	path, key, currentKey, shouldApply, err := resolveBaselineComparisonPaths(".", AnalyseRequest{BaselinePath: testBaselinePath})
-	if err != nil {
-		t.Fatalf("baseline path branch: %v", err)
-	}
-	if !shouldApply || path != testBaselinePath || key != "" {
-		t.Fatalf("unexpected baseline path resolution: path=%q key=%q shouldApply=%v", path, key, shouldApply)
-	}
-	if currentKey == "" {
-		t.Fatalf("expected current key to resolve in git repo")
-	}
-
-	path, key, currentKey, shouldApply, err = resolveBaselineComparisonPaths(".", AnalyseRequest{
-		BaselineStorePath: baselineStorePath,
-		BaselineKey:       "label:weekly",
-	})
-	if err != nil {
-		t.Fatalf("baseline store branch: %v", err)
-	}
-	if !shouldApply || key != "label:weekly" || !strings.HasSuffix(path, "label_weekly.json") {
-		t.Fatalf("unexpected baseline-store resolution: path=%q key=%q shouldApply=%v", path, key, shouldApply)
-	}
-	if currentKey == "" {
-		t.Fatalf("expected current key with baseline-store branch")
-	}
-
-	nonRepo := filepath.Join(t.TempDir(), "nonexistent", "repo")
-	if _, _, _, _, err := resolveBaselineComparisonPaths(nonRepo, AnalyseRequest{BaselineStorePath: baselineStorePath}); err == nil {
-		t.Fatalf("expected baseline-store without key in non-git dir to fail")
-	}
-}
-
-func TestResolveCurrentBaselineKeyBranches(t *testing.T) {
-	if key := resolveCurrentBaselineKey("."); !strings.HasPrefix(key, "commit:") {
-		t.Fatalf("expected commit key in repo, got %q", key)
-	}
-	nonRepo := filepath.Join(t.TempDir(), "nonexistent", "repo")
-	if key := resolveCurrentBaselineKey(nonRepo); key != "" {
-		t.Fatalf("expected empty key outside git repo, got %q", key)
-	}
-}
-
-func TestSaveBaselineIfNeededAlreadyExistsError(t *testing.T) {
-	application := &App{Formatter: report.NewFormatter()}
-	base := report.Report{SchemaVersion: "0.1.0", RepoPath: "."}
-	req := AnalyseRequest{
-		SaveBaseline:      true,
-		BaselineStorePath: t.TempDir(),
-		BaselineLabel:     "nightly",
-	}
-	if _, err := application.saveBaselineIfNeeded(base, ".", req, testTime()); err != nil {
-		t.Fatalf("first save baseline: %v", err)
-	}
-	if _, err := application.saveBaselineIfNeeded(base, ".", req, testTime()); err == nil {
-		t.Fatalf("expected immutable baseline key reuse to fail")
-	}
-}
-
-func TestExecuteAnalyseForwardsRustRecommendationThreshold(t *testing.T) {
-	analyzer := &fakeAnalyzer{
-		report: report.Report{
-			RepoPath: ".",
-			Dependencies: []report.DependencyReport{
-				{Name: "serde", Language: "rust", UsedExportsCount: 1, TotalExportsCount: 4, UsedPercent: 25},
-			},
-		},
-	}
-	application := &App{Analyzer: analyzer, Formatter: report.NewFormatter()}
-
-	req := DefaultRequest()
-	req.Mode = ModeAnalyse
-	req.Analyse.Language = "rust"
-	req.Analyse.Dependency = "serde"
-	req.Analyse.Format = report.FormatJSON
-	req.Analyse.Thresholds = thresholds.Values{
-		FailOnIncreasePercent:             0,
-		LowConfidenceWarningPercent:       35,
-		MinUsagePercentForRecommendations: 70,
-	}
-
-	if _, err := application.Execute(context.Background(), req); err != nil {
-		t.Fatalf(executeAnalyseErrFmt, err)
-	}
-	if analyzer.lastReq.MinUsagePercentForRecommendations == nil || *analyzer.lastReq.MinUsagePercentForRecommendations != 70 {
-		t.Fatalf("expected min-usage threshold to be forwarded for rust analysis, got %#v", analyzer.lastReq.MinUsagePercentForRecommendations)
-	}
-}
-
-func TestPrepareRuntimeTraceWithRuntimeCommand(t *testing.T) {
-	repo := t.TempDir()
-	req := DefaultRequest()
-	req.RepoPath = repo
-	req.Analyse.RuntimeTestCommand = "make -v"
-
-	warnings, tracePath := prepareRuntimeTrace(context.Background(), req)
-	if len(warnings) != 0 {
-		t.Fatalf("did not expect warnings from runtime planning: %#v", warnings)
-	}
-	if tracePath != "" {
-		t.Fatalf("expected runtime planning to defer default trace path resolution, got %q", tracePath)
-	}
-}
-
-func testTime() time.Time {
-	return time.Date(2026, time.February, 22, 15, 0, 0, 0, time.UTC)
-}
-
-const missingRuntimeMakeTarget = "make __missing_target__"
-
-func TestPrepareRuntimeTraceKeepsExplicitTracePath(t *testing.T) {
-	repo := t.TempDir()
-	explicitPath := filepath.Join(repo, ".artifacts", "explicit.ndjson")
-	req := DefaultRequest()
-	req.RepoPath = repo
-	req.Analyse.RuntimeTracePath = explicitPath
-	req.Analyse.RuntimeTestCommand = missingRuntimeMakeTarget
-
-	warnings, tracePath := prepareRuntimeTrace(context.Background(), req)
-	if len(warnings) != 0 {
-		t.Fatalf("expected runtime planning with explicit path to avoid warnings, got %#v", warnings)
-	}
-	if tracePath != req.Analyse.RuntimeTracePath {
-		t.Fatalf("expected explicit trace path to be retained, got %q", tracePath)
-	}
-}
-
-func TestPrepareRuntimeTraceMissingWorkingDirectoryStillReturnsConfiguredPath(t *testing.T) {
-	orphanedCWD := t.TempDir()
-	t.Chdir(orphanedCWD)
-	if err := os.RemoveAll(orphanedCWD); err != nil {
-		t.Fatalf("remove orphaned cwd: %v", err)
-	}
-
-	explicitPath := filepath.Join(t.TempDir(), testRuntimeTracePath)
-	req := DefaultRequest()
-	req.RepoPath = "."
-	req.Analyse.RuntimeTracePath = explicitPath
-	req.Analyse.RuntimeTestCommand = missingRuntimeMakeTarget
-
-	warnings, tracePath := prepareRuntimeTrace(context.Background(), req)
-	if len(warnings) != 0 {
-		t.Fatalf("expected runtime planning to avoid capture warnings, got %#v", warnings)
-	}
-	if tracePath != explicitPath {
-		t.Fatalf("expected explicit trace path to be retained, got %q", tracePath)
-	}
-}
-
-func TestAppendNotificationWarningsNilReportData(t *testing.T) {
-	application := &App{Notify: notify.NewDefaultDispatcher()}
-	cfg := notify.Config{
-		Slack: notify.ChannelConfig{
-			WebhookURL: "https://hooks.slack.com/services/A/B/SECRET",
-			Trigger:    notify.TriggerAlways,
-		},
-	}
-	application.appendNotificationWarnings(context.Background(), cfg, nil, notify.Outcome{})
-}
-
-func TestPrepareRuntimeTraceWithoutCommandUsesProvidedTracePath(t *testing.T) {
-	req := DefaultRequest()
-	req.Analyse.RuntimeTracePath = testRuntimeTracePath
-	warnings, tracePath := prepareRuntimeTrace(context.Background(), req)
-	if len(warnings) != 0 {
-		t.Fatalf("did not expect warnings without runtime command, got %#v", warnings)
-	}
-	if tracePath != testRuntimeTracePath {
-		t.Fatalf("expected provided trace path without capture command, got %q", tracePath)
-	}
-}
-
-func TestExecuteAnalyseForwardsRuntimeCaptureInputs(t *testing.T) {
-	analyzer := &fakeAnalyzer{
-		report: report.Report{
-			RepoPath: ".",
-			Dependencies: []report.DependencyReport{
-				{Name: "lodash", UsedExportsCount: 1, TotalExportsCount: 2, UsedPercent: 50},
-			},
-		},
-	}
-	application := &App{Analyzer: analyzer, Formatter: report.NewFormatter()}
-
-	req := DefaultRequest()
-	req.Mode = ModeAnalyse
-	req.RepoPath = t.TempDir()
-	req.Analyse.TopN = 1
-	req.Analyse.Format = report.FormatJSON
-	req.Analyse.RuntimeTracePath = filepath.Join(req.RepoPath, ".artifacts", "explicit.ndjson")
-	req.Analyse.RuntimeTestCommand = missingRuntimeMakeTarget
-
-	_, err := application.Execute(context.Background(), req)
-	if err != nil {
-		t.Fatalf("execute analyse forwarding runtime inputs: %v", err)
-	}
-	if analyzer.lastReq.RuntimeTestCommand != req.Analyse.RuntimeTestCommand {
-		t.Fatalf("expected runtime test command to be forwarded, got %q", analyzer.lastReq.RuntimeTestCommand)
-	}
-	if analyzer.lastReq.RuntimeTracePath != req.Analyse.RuntimeTracePath {
-		t.Fatalf("expected runtime trace path to be forwarded, got %q", analyzer.lastReq.RuntimeTracePath)
-	}
-	if !analyzer.lastReq.RuntimeTracePathExplicit {
-		t.Fatalf("expected explicit runtime trace path marker to be forwarded")
-	}
-}
-
-func TestExecuteAnalyseLockfileDriftWarnPolicy(t *testing.T) {
-	repo := t.TempDir()
-	if err := os.WriteFile(filepath.Join(repo, "package.json"), []byte("{}\n"), 0o600); err != nil {
-		t.Fatalf("write package.json: %v", err)
-	}
-	analyzer := &fakeAnalyzer{
-		report: report.Report{
-			RepoPath: ".",
-			Dependencies: []report.DependencyReport{
-				{Name: "lodash", UsedExportsCount: 1, TotalExportsCount: 2, UsedPercent: 50},
-			},
-		},
-	}
-	application := &App{Analyzer: analyzer, Formatter: report.NewFormatter()}
-
-	req := DefaultRequest()
-	req.Mode = ModeAnalyse
-	req.RepoPath = repo
-	req.Analyse.TopN = 1
-	req.Analyse.Format = report.FormatJSON
-	req.Analyse.Thresholds.LockfileDriftPolicy = "warn"
-
-	output, err := application.Execute(context.Background(), req)
-	if err != nil {
-		t.Fatalf("execute analyse with lockfile drift warn: %v", err)
-	}
-	if !strings.Contains(output, "lockfile drift detected") {
-		t.Fatalf("expected lockfile drift warning in output, got %q", output)
-	}
-}
-
-func TestExecuteAnalyseLockfileDriftFailPolicy(t *testing.T) {
-	repo := t.TempDir()
-	if err := os.WriteFile(filepath.Join(repo, "go.mod"), []byte("module example.com/demo\n\ngo 1.22\n\nrequire github.com/some/dep v1.0.0\n"), 0o600); err != nil {
-		t.Fatalf("write go.mod: %v", err)
-	}
-	analyzer := &fakeAnalyzer{
-		report: report.Report{
-			Dependencies: []report.DependencyReport{
-				{Name: "dep", UsedExportsCount: 1, TotalExportsCount: 2, UsedPercent: 50},
-			},
-		},
-	}
-	application := &App{Analyzer: analyzer, Formatter: report.NewFormatter()}
-
-	req := DefaultRequest()
-	req.Mode = ModeAnalyse
-	req.RepoPath = repo
-	req.Analyse.TopN = 1
-	req.Analyse.Thresholds.LockfileDriftPolicy = "fail"
-
-	_, err := application.Execute(context.Background(), req)
-	if !errors.Is(err, ErrLockfileDrift) {
-		t.Fatalf("expected ErrLockfileDrift, got %v", err)
-	}
-	if analyzer.called {
-		t.Fatalf("expected pre-analysis lockfile check to fail before analyzer execution")
-	}
-}
-
-func TestSaveBaselineIfNeededDisabledNoop(t *testing.T) {
-	application := &App{Formatter: report.NewFormatter()}
-	input := report.Report{RepoPath: ".", Warnings: []string{"keep"}}
-	updated, err := application.saveBaselineIfNeeded(input, ".", AnalyseRequest{}, testTime())
-	if err != nil {
-		t.Fatalf("save baseline noop: %v", err)
-	}
-	if len(updated.Warnings) != 1 || updated.Warnings[0] != "keep" {
-		t.Fatalf("expected unchanged report on noop save baseline, got %#v", updated)
-	}
-}
-
-func TestApplyBaselineIfNeededNoopWhenNoBaselineConfigured(t *testing.T) {
-	application := &App{Formatter: report.NewFormatter()}
-	input := report.Report{RepoPath: ".", Warnings: []string{"keep"}}
-
-	updated, err := application.applyBaselineIfNeeded(input, ".", AnalyseRequest{})
-	if err != nil {
-		t.Fatalf("apply baseline noop: %v", err)
-	}
-	if len(updated.Warnings) != 1 || updated.Warnings[0] != "keep" {
-		t.Fatalf("expected report to remain unchanged, got %#v", updated)
-	}
-}
-
-func TestApplyBaselineIfNeededReturnsConfigResolutionError(t *testing.T) {
-	application := &App{Formatter: report.NewFormatter()}
-	input := report.Report{RepoPath: ".", Warnings: []string{"keep"}}
-
-	_, err := application.applyBaselineIfNeeded(input, filepath.Join(t.TempDir(), "nonexistent", "repo"), AnalyseRequest{
-		BaselineStorePath: ".artifacts/baselines",
-	})
-	if err == nil {
-		t.Fatalf("expected baseline config resolution error")
-	}
-}
-
-func TestExecuteAnalyseReturnsFormattedOutputWhenSaveBaselineValidationFails(t *testing.T) {
-	analyzer := &fakeAnalyzer{
-		report: report.Report{
-			RepoPath:      ".",
-			Dependencies:  []report.DependencyReport{{Name: "dep", UsedExportsCount: 1, TotalExportsCount: 2, UsedPercent: 50}},
-			SchemaVersion: "0.1.0",
-		},
-	}
-	application := &App{Analyzer: analyzer, Formatter: report.NewFormatter()}
-
-	req := DefaultRequest()
-	req.Mode = ModeAnalyse
-	req.Analyse.TopN = 1
-	req.Analyse.Format = report.FormatJSON
-	req.Analyse.SaveBaseline = true
-
-	output, err := application.Execute(context.Background(), req)
-	if err == nil {
-		t.Fatalf("expected save-baseline validation error")
-	}
-	if !strings.Contains(err.Error(), saveBaselineStoreErr) {
-		t.Fatalf("unexpected save-baseline error: %v", err)
-	}
-	if !strings.Contains(output, "\"dependencies\"") {
-		t.Fatalf("expected formatted output to be returned alongside error, got %q", output)
-	}
-}
-
-func TestExecuteAnalyseReturnsFormatterErrorWhenNoPriorError(t *testing.T) {
-	analyzer := &fakeAnalyzer{
-		report: report.Report{
-			RepoPath:     ".",
-			Dependencies: []report.DependencyReport{{Name: "dep", UsedExportsCount: 1, TotalExportsCount: 2, UsedPercent: 50}},
-		},
-	}
-	application := &App{Analyzer: analyzer, Formatter: report.NewFormatter()}
-
-	req := DefaultRequest()
-	req.Mode = ModeAnalyse
-	req.Analyse.TopN = 1
-	req.Analyse.Format = report.Format("invalid")
-
-	_, err := application.Execute(context.Background(), req)
-	if err == nil || !strings.Contains(strings.ToLower(err.Error()), "unknown format") {
-		t.Fatalf("expected formatter error, got %v", err)
-	}
-}
-
-func TestExecuteAnalyseApplyBaselineErrorPreservesOriginalWhenFormatFails(t *testing.T) {
-	analyzer := &fakeAnalyzer{
-		report: report.Report{
-			RepoPath: ".",
-			Dependencies: []report.DependencyReport{
-				{Name: "dep", UsedExportsCount: 1, TotalExportsCount: 2, UsedPercent: 50},
-			},
-		},
-	}
-	application := &App{Analyzer: analyzer, Formatter: report.NewFormatter()}
-	req := DefaultRequest()
-	req.Mode = ModeAnalyse
-	req.Analyse.TopN = 1
-	req.Analyse.Format = report.Format("invalid")
-	req.Analyse.BaselinePath = filepath.Join(t.TempDir(), missingBaselineFileName)
-
-	_, err := application.Execute(context.Background(), req)
-	if err == nil {
-		t.Fatalf("expected apply-baseline error")
-	}
-	if strings.Contains(strings.ToLower(err.Error()), "unknown format") {
-		t.Fatalf("expected original baseline error, got %v", err)
-	}
-}
-
-func TestExecuteAnalyseFailOnIncreasePreservesOriginalWhenFormatFails(t *testing.T) {
-	delta := 5.0
-	analyzer := &fakeAnalyzer{
-		report: report.Report{
-			RepoPath:             ".",
-			WasteIncreasePercent: &delta,
-			Dependencies:         []report.DependencyReport{{Name: "dep", UsedExportsCount: 1, TotalExportsCount: 2, UsedPercent: 50}},
-		},
-	}
-	application := &App{Analyzer: analyzer, Formatter: report.NewFormatter()}
-	req := DefaultRequest()
-	req.Mode = ModeAnalyse
-	req.Analyse.TopN = 1
-	req.Analyse.Format = report.Format("invalid")
-	req.Analyse.Thresholds.FailOnIncreasePercent = 1
-
-	_, err := application.Execute(context.Background(), req)
-	if !errors.Is(err, ErrFailOnIncrease) {
-		t.Fatalf("expected ErrFailOnIncrease, got %v", err)
-	}
-}
-
-func TestExecuteAnalyseSaveBaselineErrorPreservesOriginalWhenFormatFails(t *testing.T) {
-	analyzer := &fakeAnalyzer{
-		report: report.Report{
-			RepoPath: ".",
-			Dependencies: []report.DependencyReport{
-				{Name: "dep", UsedExportsCount: 1, TotalExportsCount: 2, UsedPercent: 50},
-			},
-		},
-	}
-	application := &App{Analyzer: analyzer, Formatter: report.NewFormatter()}
-	req := DefaultRequest()
-	req.Mode = ModeAnalyse
-	req.Analyse.TopN = 1
-	req.Analyse.Format = report.Format("invalid")
-	req.Analyse.SaveBaseline = true
-
-	_, err := application.Execute(context.Background(), req)
-	if err == nil {
-		t.Fatalf("expected save-baseline error")
-	}
-	if !strings.Contains(err.Error(), saveBaselineStoreErr) {
-		t.Fatalf("expected save-baseline store error, got %v", err)
-	}
-}
